feat: complete task for modules
