+++ src/modules/esm.mjs
@@ -1,0 +1,46 @@
+import { dirname, join, sep } from 'node:path';
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'node:http';
+import { readFile } from 'node:fs/promises';
+import { fileURLToPath } from 'node:url';
+import './files/c.cjs';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const random = Math.random();
+const jsonPath = join(__dirname, 'files', random > 0.5 ? 'a.json' : 'b.json');
+
+let unknownObject;
+
+try {
+    unknownObject = JSON.parse(await readFile(jsonPath, 'utf-8'));
+} catch (error) {
+    console.error('Failed to parse JSON:', error);
+    unknownObject = { error: 'Invalid JSON' };
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export default {
+    unknownObject,
+    myServer,
+};
